+++ app/src/main/java/com/fibelatti/photowidget/platform/DialogKtx.kt
@@ -1,5 +1,6 @@
 package com.fibelatti.photowidget.platform
 
+import android.content.Context
 import androidx.appcompat.app.AppCompatDialog
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.ViewModelStoreOwner
@@ -7,6 +8,11 @@ import androidx.lifecycle.setViewTreeLifecycleOwner
 import androidx.lifecycle.setViewTreeViewModelStoreOwner
 import androidx.savedstate.SavedStateRegistryOwner
 import androidx.savedstate.setViewTreeSavedStateRegistryOwner
+import com.fibelatti.photowidget.R
+import com.fibelatti.photowidget.di.PhotoWidgetEntryPoint
+import com.fibelatti.photowidget.di.entryPoint
+import com.google.android.material.color.DynamicColors
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 
 fun AppCompatDialog.setViewTreeOwners() {
     val activity = context.findActivity()
@@ -19,4 +25,22 @@ fun AppCompatDialog.setViewTreeOwners() {
         decorView.setViewTreeViewModelStoreOwner(activity as? ViewModelStoreOwner)
         decorView.setViewTreeSavedStateRegistryOwner(activity as? SavedStateRegistryOwner)
     }
+}
+
+fun Context.showMaterialAlertDialog(body: MaterialAlertDialogBuilder.() -> Unit) {
+    val entryPoint: PhotoWidgetEntryPoint = entryPoint(this)
+
+    val builder: MaterialAlertDialogBuilder = if (entryPoint.userPreferencesStorage().dynamicColors) {
+        MaterialAlertDialogBuilder(
+            /* context = */ DynamicColors.wrapContextIfAvailable(this),
+            /* overrideThemeResId = */ R.style.AppTheme_MaterialDialog,
+        )
+    } else {
+        MaterialAlertDialogBuilder(
+            /* context = */ this,
+            /* overrideThemeResId = */ R.style.AppTheme_MaterialDialog_DefaultColors,
+        )
+    }
+
+    builder.apply { body() }.show()
 }